archive/analisi_dati/requirements.txt
# Pacchetti Python per analisi GCV

# Astronomia
astropy>=5.3
astroquery>=0.4.6
healpy>=1.16.0

# Calcolo scientifico
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# GPU computing
cupy-cuda12x>=12.0.0  # Per CUDA 12.x, adatta alla tua versione
numba>=0.57.0

# Plotting
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine learning (opzionale per fit avanzati)
scikit-learn>=1.3.0

# I/O e utilità
h5py>=3.9.0
tqdm>=4.65.0
requests>=2.31.0

# Analisi statistica
emcee>=3.1.0  # MCMC per fit bayesiani

# Cataloghi astronomici
astro-prospector  # Per accesso SDSS
